refactor: extract module
Extract table formatting logic to own module. This way other output formatting functions can be grouped together an reused.
